Genai assistant completed with tests in core modules (#11)
* Genai assistant added on pagila queries
* Genai assistant merged in root angular module
* New views, APIs and models created
* New views, APIs and models created: stage 2
* New views, APIs and models created: stage 3
* New views, APIs and models created: stage 4
* wire and add get_store_comparison,get_revenue_summary,get_slow_moving_films MCP tools: stage 5
* langgraph improves with summary and validation logic: stage 6
* complete tests across modules: stage 7

+++ data-ebean/src/main/java/com/github/brane08/pagila/film/entities/Film.java
@@ -6,6 +6,7 @@ import com.github.brane08.pagila.seedworks.entities.YearConverter;
 import io.ebean.annotation.DbArray;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.Year;
 import java.util.List;
 
@@ -40,13 +41,13 @@ public class Film extends BaseModel {
     Integer rentalDuration;
 
     @Column(name = "rental_rate")
-    Double rentalRate;
+    BigDecimal rentalRate;
 
     @Column(name = "length", columnDefinition = "int2")
     Integer length;
 
     @Column(name = "replacement_cost")
-    Double replacementCost;
+    BigDecimal replacementCost;
 
     @Column(name = "rating_txt", columnDefinition = "varchar(10)")
     FilmRating rating;
@@ -125,11 +126,11 @@ public class Film extends BaseModel {
         this.rentalDuration = rentalDuration;
     }
 
-    public Double getRentalRate() {
+    public BigDecimal getRentalRate() {
         return rentalRate;
     }
 
-    public void setRentalRate(Double rentalRate) {
+    public void setRentalRate(BigDecimal rentalRate) {
         this.rentalRate = rentalRate;
     }
 
@@ -141,11 +142,11 @@ public class Film extends BaseModel {
         this.length = length;
     }
 
-    public Double getReplacementCost() {
+    public BigDecimal getReplacementCost() {
         return replacementCost;
     }
 
-    public void setReplacementCost(Double replacementCost) {
+    public void setReplacementCost(BigDecimal replacementCost) {
         this.replacementCost = replacementCost;
     }
